All renders correctly. I'm just deciding if I need to change my component structure so I can begin styling.
