+++ starwars/src/App.js
@@ -2,6 +2,7 @@ import React from 'react';
 import './App.css';
 
 import Card from './components/CardComponents/Card.js';
+import Header from "./components/HeaderComponents/Header.js";
 
 const App = () => {
   // Try to think through what state you'll need for this app before starting. Then build out
@@ -13,7 +14,7 @@ const App = () => {
 
   return (
     <div className="App">
-      <h1 className="Header">React Wars</h1>
+      <Header />
       <Card />
     </div>
   );
