+++ starwars/src/components/CardComponents/Card.css
@@ -1,0 +1,3 @@
+.card {
+   border: 3px solid white;
+}
