+++ starwars/src/components/CardComponents/Card.js
@@ -1,20 +1,35 @@
 import React, { useState, useEffect } from 'react';
 import axios from 'axios';
-
-import Data from './DataComponents/Data.js';
-import CharImg from './ImgComponents/CharImg.js';
-import PlanImg from './ImgComponents/PlanImg.js';
+import "./Card.css";
 
 function Card() {
 
+  const [bio, setBio] = useState([]);
+
+  useEffect(() => {
+    axios.get("https://swapi.co/api/people/")
+      .then(res => {
+        console.log(res.data.results);
+        setBio(res.data.results);
+      })
+      .catch(res => {
+        console.log("Error in retrieving data: ", res);
+      })
+  }, [])
+
   return (
-    return (
-      <div className="card">
-        <Portrait />
-        <Bio />
-        <PlanImg />
-      </div>
-    )
+    <div>
+      {bio.map((info) => {
+        console.log(info);
+        return (
+          <>
+            <p><strong>Name:</strong>{info.name}</p>
+            <p><strong>Birth Year:</strong>{info.birth_year}</p>
+          </>
+        )
+      })}
+
+    </div>
   )
 };
