+++ starwars/src/components/HeaderComponents/Header.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import "./Header.css";
+
+function Header() {
+  return (
+    <>
+      <h1>React Wars</h1>
+    </>
+  )
+}
+
+export default Header;
